Some fix

+++ bin/start.sh
@@ -12,9 +12,11 @@ mkdir $DB_PATH
 mongod --dbpath="$DB_PATH" --port="$DB_PORT" &
 PS1=$!
 
-# start client
-$LOCAL_BIN/webpack-dev-server --progress --colors --hot --inline --watch &
-PS2=$!
+# start client developing
+if [[ NODE_ENV != "prouction" ]]; then
+  $LOCAL_BIN/webpack-dev-server --progress --colors --hot --inline --watch &
+  PS2=$!
+fi
 
 # wait mongo start
 COUNTER=0
@@ -34,9 +36,14 @@ done
 # migrate
 $LOCAL_BIN/babel-node ./server/migrate/index.js
 
-# start server
-$LOCAL_BIN/nodemon --exec "$LOCAL_BIN/babel-node" -- ./server/index.js
-
-# kill all related process
-kill -9 $PS1
-kill -9 $PS2
+if [[ NODE_ENV == 'production' ]]; then
+  PORT=3000
+  # start server as a process
+  PORT=3001 $LOCAL_BIN/forever start "node ./lib/index.js"
+else
+  # start dev server
+  $LOCAL_BIN/nodemon --exec "$LOCAL_BIN/babel-node" -- ./server/index.js
+  # kill all related process
+  kill -9 $PS1
+  kill -9 $PS2
+fi
